refactor: rename object_request to storage_request, atom to expr, remove unnecessary type coercions
- Rename object_request/object_response communication modules to storage_request/storage_response for consistency with the storage layer naming
- Rename atom variables/identifiers to expr in cold storage collation, indexing, and merging
- Remove redundant bytes() and int() wrapping in models, validation, and consensus code
- Restructure Block and Receipt serialization to embed version in the body link chain rather than a nested header
- Update config keys expr_fetch_interval/expr_fetch_retries to storage_fetch_interval/storage_fetch_retries
- Update tests to match renamed modules and simplified type handling

+++ src/astreum/storage/thread.py
@@ -53,7 +53,7 @@ def _update_storage_request_price(node: "Node") -> None:
     pressure = _incoming_queue_pressure(node)
     factor = _price_movement_factor_for_pressure(pressure)
     previous_price = node.storage_request_current_price
-    target_price = max(node.config["storage_request_minimum_price"], int(round(previous_price * factor)))
+    target_price = max(node.config["storage_request_minimum_price"], round(previous_price * factor))
 
     
     node.storage_request_current_price = target_price
